Convert CHANGES and README to Markdown, unify history
In particular, remove the separate changes history from the manual, so that
now the CHANGES (resp. CHANGES.md) file is the unique reference for the
history of the package. That means we only need to change one place for
updates, and also people only have to check a single place to find out what
has changed.

Also update some statements in the README and point to the GitHub issue
tracker.